fix(roller): audit remediation X3/L4/L3/L7
X3 — Linearize O(n^2) reroll/unique accumulators. Both modifiers folded
over the pool with reduce + spread-copy of the result array each iteration
(and unique cloned its dedup Set per element), giving O(n^2) copying for
pools up to MAX_QUANTITY (10000). Rewrote to single-pass push into one
array with one persistent Set, preserving exact behavior. Measured on
5000d10000U: ~59ms -> ~2ms (~29x); 1000d6R{1}: ~1.04ms -> ~0.24ms (~4.3x).
Added large-pool benchmark cases (1000d6, 100d6R{<3}, 5000d10000U).

L4 — ReDoS hardening. Added bounded-execution property test feeding random,
unicode, and adversarial notation strings to notation()/isDiceNotation(),
asserting each call stays well under a 50ms budget. LRU parse cache deferred
(no batch path yet). Description-path face allocation already symbolic in
current optionsToSidesFaces; finding describes stale code.

L7a — Split the 1078-LOC ModifierEngine.test.ts into apply / format /
parseValidate files, all <600 LOC, no test loss (148 tests preserved).

L3a — Table-driven parseArguments dispatch. Replaced the special-die
if-ladder with a SPECIAL_DIE_PARSERS table, removing the three unsafe
as RollParams<T>[] casts by invoking each parser with the caller's T.
