+++ packages/roller/__tests__/lib/modifiers/ModifierEngine.apply.test.ts
@@ -1,14 +1,6 @@
 import { describe, expect, test } from 'bun:test'
 import type { ModifierContext } from '../../../src/modifiers/schema'
-import {
-  applyAllModifiers,
-  applyModifier,
-  getModifier,
-  modifierToDescription,
-  modifierToNotation,
-  validateModifiers
-} from '../../../src/modifiers/registry'
-import { parseModifiers } from '../../../src/notation/parse/parseModifiers'
+import { applyAllModifiers, applyModifier, getModifier } from '../../../src/modifiers/registry'
 
 const mockRollOne = (): number => Math.floor(Math.random() * 6) + 1
 const mockContext: ModifierContext = {
@@ -413,280 +405,6 @@ describe('applyModifier', () => {
   })
 })
 
-describe('modifierToDescription', () => {
-  test('generates description for plus modifier', () => {
-    expect(modifierToDescription('plus', 5)).toEqual(['Add 5'])
-  })
-
-  test('generates description for minus modifier', () => {
-    expect(modifierToDescription('minus', 3)).toEqual(['Subtract 3'])
-  })
-
-  test('generates description for cap modifier', () => {
-    const result = modifierToDescription('cap', { greaterThan: 5 })
-    expect(result).toEqual(['No Rolls greater than 5'])
-  })
-
-  test('generates description for drop modifier', () => {
-    const result = modifierToDescription('drop', { lowest: 1 })
-    expect(result).toEqual(['Drop lowest'])
-  })
-
-  test('generates description for keep highest modifier', () => {
-    const result = modifierToDescription('keep', { highest: 3 })
-    expect(result).toEqual(['Keep highest 3'])
-  })
-
-  test('generates description for keep highest 1 modifier', () => {
-    const result = modifierToDescription('keep', { highest: 1 })
-    expect(result).toEqual(['Keep highest'])
-  })
-
-  test('generates description for keep lowest modifier', () => {
-    const result = modifierToDescription('keep', { lowest: 2 })
-    expect(result).toEqual(['Keep lowest 2'])
-  })
-
-  test('generates description for keep lowest 1 modifier', () => {
-    const result = modifierToDescription('keep', { lowest: 1 })
-    expect(result).toEqual(['Keep lowest'])
-  })
-
-  test('generates description for reroll modifier', () => {
-    const result = modifierToDescription('reroll', { exact: [1, 2] })
-    expect(result).toEqual(['Reroll 1 and 2'])
-  })
-
-  test('generates description for explode modifier', () => {
-    const result = modifierToDescription('explode', true)
-    expect(result).toEqual(['Exploding Dice'])
-  })
-
-  test('generates description for unique modifier', () => {
-    const result = modifierToDescription('unique', true)
-    expect(result).toEqual(['No Duplicate Rolls'])
-  })
-
-  test('generates description for unique modifier with exceptions', () => {
-    const result = modifierToDescription('unique', { notUnique: [1, 2] })
-    expect(result).toEqual(['No Duplicates (except 1 and 2)'])
-  })
-
-  test('generates description for replace modifier', () => {
-    const result = modifierToDescription('replace', { from: 1, to: 6 })
-    expect(result).toEqual(['Replace 1 with 6'])
-  })
-
-  test('returns undefined for undefined options', () => {
-    const result = modifierToDescription('plus', undefined)
-    expect(result).toBeUndefined()
-  })
-
-  test('generates description for compound modifier', () => {
-    expect(modifierToDescription('compound', true)).toEqual(['Compounding Dice'])
-  })
-
-  test('generates description for compound modifier with depth', () => {
-    expect(modifierToDescription('compound', 3)).toEqual(['Compounding Dice (max 3 times)'])
-  })
-
-  test('generates description for compound modifier unlimited', () => {
-    expect(modifierToDescription('compound', 0)).toEqual(['Compounding Dice (unlimited)'])
-  })
-
-  test('generates description for penetrate modifier', () => {
-    expect(modifierToDescription('penetrate', true)).toEqual(['Penetrating Dice'])
-  })
-
-  test('generates description for penetrate modifier with depth', () => {
-    expect(modifierToDescription('penetrate', 3)).toEqual(['Penetrating Dice (max 3 times)'])
-  })
-
-  test('generates description for penetrate modifier unlimited', () => {
-    expect(modifierToDescription('penetrate', 0)).toEqual(['Penetrating Dice (unlimited)'])
-  })
-})
-
-describe('modifierToNotation', () => {
-  test('generates notation for plus modifier', () => {
-    const result = modifierToNotation('plus', 5)
-    expect(result).toBe('+5')
-  })
-
-  test('handles negative plus modifier', () => {
-    const result = modifierToNotation('plus', -3)
-    expect(result).toBe('-3')
-  })
-
-  test('generates notation for minus modifier', () => {
-    const result = modifierToNotation('minus', 3)
-    expect(result).toBe('-3')
-  })
-
-  test('generates notation for cap modifier', () => {
-    const result = modifierToNotation('cap', { greaterThan: 5 })
-    expect(result).toBe('C{>5}')
-  })
-
-  test('generates notation for drop modifier', () => {
-    const result = modifierToNotation('drop', { lowest: 1 })
-    expect(result).toBe('L')
-  })
-
-  test('generates notation for keep highest modifier', () => {
-    const result = modifierToNotation('keep', { highest: 3 })
-    expect(result).toBe('K3')
-  })
-
-  test('generates notation for keep highest 1 modifier', () => {
-    const result = modifierToNotation('keep', { highest: 1 })
-    expect(result).toBe('K')
-  })
-
-  test('generates notation for keep lowest modifier', () => {
-    const result = modifierToNotation('keep', { lowest: 2 })
-    expect(result).toBe('kl2')
-  })
-
-  test('generates notation for keep lowest 1 modifier', () => {
-    const result = modifierToNotation('keep', { lowest: 1 })
-    expect(result).toBe('kl')
-  })
-
-  test('generates notation for reroll modifier', () => {
-    const result = modifierToNotation('reroll', { exact: [1, 2] })
-    expect(result).toBe('R{1,2}')
-  })
-
-  test('generates notation for explode modifier', () => {
-    const result = modifierToNotation('explode', true)
-    expect(result).toBe('!')
-  })
-
-  test('returns undefined for explode modifier when false', () => {
-    const result = modifierToNotation('explode', false)
-    expect(result).toBeUndefined()
-  })
-
-  test('generates notation for unique modifier', () => {
-    const result = modifierToNotation('unique', true)
-    expect(result).toBe('U')
-  })
-
-  test('generates notation for unique modifier with notUnique array', () => {
-    const result = modifierToNotation('unique', { notUnique: [1, 6] })
-    expect(result).toBe('U{1,6}')
-  })
-
-  test('returns undefined for unique modifier when false', () => {
-    const result = modifierToNotation('unique', false)
-    expect(result).toBeUndefined()
-  })
-
-  test('generates notation for replace modifier', () => {
-    const result = modifierToNotation('replace', { from: 1, to: 6 })
-    expect(result).toBe('V{1=6}')
-  })
-
-  test('returns undefined for undefined options', () => {
-    const result = modifierToNotation('plus', undefined)
-    expect(result).toBeUndefined()
-  })
-
-  test('generates notation for compound modifier', () => {
-    const result = modifierToNotation('compound', true)
-    expect(result).toBe('!!')
-  })
-
-  test('generates notation for compound modifier with depth', () => {
-    const result = modifierToNotation('compound', 3)
-    expect(result).toBe('!!3')
-  })
-
-  test('generates notation for compound modifier unlimited', () => {
-    const result = modifierToNotation('compound', 0)
-    expect(result).toBe('!!0')
-  })
-
-  test('generates notation for penetrate modifier', () => {
-    const result = modifierToNotation('penetrate', true)
-    expect(result).toBe('!p')
-  })
-
-  test('generates notation for penetrate modifier with depth', () => {
-    const result = modifierToNotation('penetrate', 3)
-    expect(result).toBe('!p3')
-  })
-
-  test('generates notation for penetrate modifier unlimited', () => {
-    const result = modifierToNotation('penetrate', 0)
-    expect(result).toBe('!p0')
-  })
-
-  test('generates notation for multiply modifier', () => {
-    const result = modifierToNotation('multiply', 2)
-    expect(result).toBe('*2')
-  })
-
-  test('generates notation for multiplyTotal modifier', () => {
-    const result = modifierToNotation('multiplyTotal', 3)
-    expect(result).toBe('**3')
-  })
-
-  test('generates notation for count modifier', () => {
-    const result = modifierToNotation('count', { greaterThanOrEqual: 7 })
-    expect(result).toBe('#{>=7}')
-  })
-
-  test('generates notation for count modifier with deduct', () => {
-    const result = modifierToNotation('count', {
-      greaterThanOrEqual: 7,
-      lessThanOrEqual: 1,
-      deduct: true
-    })
-    expect(result).toBe('#{>=7,<=1}')
-  })
-})
-
-describe('modifierToDescription', () => {
-  test('generates description for explode modifier', () => {
-    const result = modifierToDescription('explode', true)
-    expect(result).toContain('Exploding Dice')
-  })
-
-  test('returns empty array for explode when false', () => {
-    const result = modifierToDescription('explode', false)
-    expect(result).toEqual([])
-  })
-
-  test('generates description for unique modifier', () => {
-    const result = modifierToDescription('unique', true)
-    expect(result).toContain('No Duplicate Rolls')
-  })
-
-  test('generates description for unique with notUnique array', () => {
-    const result = modifierToDescription('unique', { notUnique: [1] })!
-    expect(result[0]).toContain('No Duplicates')
-    expect(result[0]).toContain('except')
-  })
-
-  test('generates description for drop lowest', () => {
-    const result = modifierToDescription('drop', { lowest: 1 })!
-    expect(result[0]).toContain('Drop')
-    expect(result[0]).toContain('lowest')
-  })
-
-  test('returns undefined result for undefined options', () => {
-    const result = modifierToDescription('plus', undefined)
-    expect(result).toBeUndefined()
-  })
-
-  test('returns empty array for unique when false', () => {
-    const result = modifierToDescription('unique', false)
-    expect(result).toEqual([])
-  })
-})
-
 describe('direct modifier definition edge cases', () => {
   describe('explode modifier', () => {
     test('parse returns empty object when notation does not match', () => {
@@ -822,128 +540,6 @@ describe('applyModifier - count modifier', () => {
   })
 })
 
-describe('modifierToDescription - additional modifiers', () => {
-  test('generates description for multiply modifier', () => {
-    const result = modifierToDescription('multiply', 2)
-    expect(result).toEqual(['Multiply dice by 2'])
-  })
-
-  test('generates description for multiplyTotal modifier', () => {
-    const result = modifierToDescription('multiplyTotal', 3)
-    expect(result).toEqual(['Multiply total by 3'])
-  })
-
-  test('generates description for count modifier', () => {
-    const result = modifierToDescription('count', { greaterThanOrEqual: 7 })
-    expect(result).toEqual(['Count dice greater than or equal to 7'])
-  })
-
-  test('generates description for count modifier with deduct', () => {
-    const result = modifierToDescription('count', {
-      greaterThanOrEqual: 7,
-      lessThanOrEqual: 1,
-      deduct: true
-    })
-    expect(result).toEqual([
-      'Count dice greater than or equal to 7, deduct dice less than or equal to 1'
-    ])
-  })
-})
-
-describe('parseModifiers', () => {
-  test('parses plus modifier from notation', () => {
-    const result = parseModifiers('2d6+5')
-    expect(result.plus).toBe(5)
-  })
-
-  test('parses minus modifier from notation', () => {
-    const result = parseModifiers('2d6-3')
-    expect(result.minus).toBe(3)
-  })
-
-  test('parses drop lowest from notation', () => {
-    const result = parseModifiers('4d6L')
-    expect(result.drop).toEqual({ lowest: 1 })
-  })
-
-  test('parses drop highest from notation', () => {
-    const result = parseModifiers('2d20H')
-    expect(result.drop).toEqual({ highest: 1 })
-  })
-
-  test('parses keep highest from notation', () => {
-    const result = parseModifiers('2d20K')
-    expect(result.keep).toEqual({ highest: 1 })
-  })
-
-  test('parses explode from notation', () => {
-    const result = parseModifiers('3d6!')
-    expect(result.explode).toBe(true)
-  })
-
-  test('parses unique from notation', () => {
-    const result = parseModifiers('4d6U')
-    expect(result.unique).toBe(true)
-  })
-
-  test('parses reroll from notation', () => {
-    const result = parseModifiers('4d6R{1}')
-    expect(result.reroll).toEqual({ exact: [1] })
-  })
-
-  test('parses cap from notation', () => {
-    const result = parseModifiers('3d6C{>5}')
-    expect(result.cap).toEqual({ greaterThan: 5 })
-  })
-
-  test('parses replace from notation', () => {
-    const result = parseModifiers('4d6V{1=6}')
-    // Replace modifier returns an array of rules
-    expect(result.replace).toEqual([{ from: 1, to: 6 }])
-  })
-
-  test('parses compound from notation', () => {
-    const result = parseModifiers('3d6!!')
-    expect(result.compound).toBe(true)
-  })
-
-  test('parses penetrate from notation', () => {
-    const result = parseModifiers('3d6!p')
-    expect(result.penetrate).toBe(true)
-  })
-
-  test('parses multiply from notation', () => {
-    const result = parseModifiers('2d6*2')
-    expect(result.multiply).toBe(2)
-  })
-
-  test('parses multiplyTotal from notation', () => {
-    const result = parseModifiers('2d6**3')
-    expect(result.multiplyTotal).toBe(3)
-  })
-
-  test('parses count from notation (#{} syntax)', () => {
-    const result = parseModifiers('5d10#{>=7}')
-    expect(result.count).toEqual({ greaterThanOrEqual: 7 })
-  })
-
-  test('parses count with deduct from notation', () => {
-    const result = parseModifiers('5d10#{>=7,<=1}')
-    expect(result.count).toEqual({ greaterThanOrEqual: 7, lessThanOrEqual: 1, deduct: true })
-  })
-
-  test('parses multiple modifiers from notation', () => {
-    const result = parseModifiers('4d6L+5')
-    expect(result.drop).toEqual({ lowest: 1 })
-    expect(result.plus).toBe(5)
-  })
-
-  test('returns empty object for notation without modifiers', () => {
-    const result = parseModifiers('2d6')
-    expect(Object.keys(result).length).toBe(0)
-  })
-})
-
 describe('applyAllModifiers', () => {
   test('applies multiple modifiers in priority order', () => {
     const fixedRollOne = (): number => 5
@@ -994,85 +590,5 @@ describe('applyAllModifiers', () => {
     expect(result.rolls).toEqual([1, 2, 3])
     expect(result.logs.length).toBe(0)
     expect(result.totalTransformers.length).toBe(0)
-  })
-})
-
-describe('validateModifiers', () => {
-  test('validates drop modifier - throws when dropping all dice', () => {
-    expect(() => {
-      validateModifiers({ drop: { lowest: 4 } }, { sides: 6, quantity: 4 })
-    }).toThrow('Cannot drop 4 dice from a pool of 4')
-  })
-
-  test('validates drop modifier - throws when dropping more than available', () => {
-    expect(() => {
-      validateModifiers({ drop: { highest: 3, lowest: 2 } }, { sides: 6, quantity: 4 })
-    }).toThrow('Cannot drop 5 dice from a pool of 4')
-  })
-
-  test('validates drop modifier - passes for valid drop count', () => {
-    expect(() => {
-      validateModifiers({ drop: { lowest: 1 } }, { sides: 6, quantity: 4 })
-    }).not.toThrow()
-  })
-
-  test('validates unique modifier - throws when quantity exceeds sides', () => {
-    expect(() => {
-      validateModifiers({ unique: true }, { sides: 4, quantity: 5 })
-    }).toThrow('Cannot have 5 unique values with only 4 sides')
-  })
-
-  test('validates unique modifier - passes when quantity equals sides', () => {
-    expect(() => {
-      validateModifiers({ unique: true }, { sides: 6, quantity: 6 })
-    }).not.toThrow()
-  })
-
-  test('validates unique modifier - passes when quantity less than sides', () => {
-    expect(() => {
-      validateModifiers({ unique: true }, { sides: 6, quantity: 4 })
-    }).not.toThrow()
-  })
-
-  test('validates keep modifier - throws when keeping more highest than available', () => {
-    expect(() => {
-      validateModifiers({ keep: { highest: 5 } }, { sides: 6, quantity: 4 })
-    }).toThrow('Cannot keep 5 highest dice from a pool of 4')
-  })
-
-  test('validates keep modifier - throws when keeping less than 1 highest', () => {
-    expect(() => {
-      validateModifiers({ keep: { highest: 0 } }, { sides: 6, quantity: 4 })
-    }).toThrow('Cannot keep 0 highest dice from a pool of 4')
-  })
-
-  test('validates keep modifier - throws when keeping more lowest than available', () => {
-    expect(() => {
-      validateModifiers({ keep: { lowest: 5 } }, { sides: 6, quantity: 4 })
-    }).toThrow('Cannot keep 5 lowest dice from a pool of 4')
-  })
-
-  test('validates keep modifier - throws when keeping less than 1 lowest', () => {
-    expect(() => {
-      validateModifiers({ keep: { lowest: 0 } }, { sides: 6, quantity: 4 })
-    }).toThrow('Cannot keep 0 lowest dice from a pool of 4')
-  })
-
-  test('validates keep modifier - passes for valid keep count', () => {
-    expect(() => {
-      validateModifiers({ keep: { highest: 3 } }, { sides: 6, quantity: 4 })
-    }).not.toThrow()
-  })
-
-  test('validates multiple modifiers together', () => {
-    expect(() => {
-      validateModifiers({ drop: { lowest: 1 }, unique: true }, { sides: 6, quantity: 4 })
-    }).not.toThrow()
-  })
-
-  test('handles modifiers without validate function', () => {
-    expect(() => {
-      validateModifiers({ plus: 5, minus: 3 }, { sides: 6, quantity: 2 })
-    }).not.toThrow()
   })
 })
